Improved live updates on decoder view.
Now when moving or deleting fields in the decoder view, the current index, the "First row decoded" column and the "Decoded as XXX" text entries are updated accordingly to reflect the changes in the TreeView.

That means you no longer need to exit the decoder view to see if the schema is correct, speeding up decoding tables a lot.