Now "Save as" initial folder is more consistent.
If you try to save a new packfile, it'll use by default the data folder
of the currently selected game.
If you try to save an existing PackFile, it'll use by default the
current folder of the original Packfile.

Also, as a consecuence, the GameSelected struct has been updated too to
actually be somewhat useful.