Gigamega DB & Decoder patch.
This was supposed to be a little patch to add list support for the DB
tables. One thing led to another and I had to rework the entire DB logic
to make it work with lists. It didn't worked. So I removed all the
list-related code, and decided to at least move all the DB stuff out of
the `main.rs` file (around 2000 lines). That worked.

Also, the specific decoders have been changed to just take a mutable
reference to index, instead of receiving it, creating another and
returning it. Because, well, I was already tweaking the DB code so...
why not doing this too?

So more or less, that's the reason this patch is so big.

PD: Also, 3 helper functions have been moved from `main.rs` to
`ui/mod.rs`, because reasons (cleaning, mainly).