Moved from CSV to TSV files for Export/Import DB Tables and Loc Files.
The reason is Excel and Libreoffice both break in different ways the CSV
files when saving them. They don't break TSV files.

Also the checks when trying to import TSV files have been improved. Now
RPFM can easely check if a TSV belongs to a Table/Version or a Loc.