Moved all the Loc stuff out of "main.rs".
Also, completely revised the Copy/Paste stuff for both cells and rows,
and both DB Tables and Loc files. Now it's faaaar more stable, works
better with Excel/Libreoffice, and it checks the lines can be pasted
before even attempting to paste them.

Also, fixed the non-working "Paste rows" for Loc files. A check that was
done before it was needed was the culprit.

And fixed a CTD when copying the dummy column.