Macro patch to use paths instead of indexes for PackedFiles.
Explanation: recently, a bug was found that caused CTD to random people (not to everyone, because fuck consistency) after starting the game when the PackedFiles inside a PackFile were not sorted alphabetically. The fix was to sort them on save. The problem is that sorting them changes the index of the PackedFiles and, if we have a PackedFile open (for example, a table), his index will be wrong after saving, and it'll start saving data in the wrong PackedFile.

The solution: stop using the index in the vector to identify the tables, and use their path instead. But that has required rewriting a ton of code.

Good thing we got from this is that we can finally identify what PackedFile is open and we can allow to delete/rename any other PackedFile while it's open. Also, a few instances were PackedFiles were saved twice after an edit have been fixed.

Still, there is a bug where you can add a PackedFile overwriting an existing one while it's open, but that's a different bug.