Improved type detection in DB Table edits.
Now you can't insert invalid data in a column anymore other than manually and, if you do it, the change will be reversed. Also, more cleanups.

NOTE: for some reason after doing the undo the table becomes uneditable for a few seconds. That's something to fix later, as I can't come up for a cause for that.